+++ src/Underscore/Types/Arrays.php
@@ -11,6 +11,12 @@ use \Underscore\Traits\Repository;
 class Arrays extends Repository
 {
   /**
+   * The method used to convert new subjects
+   * @var string
+   */
+  protected $typecaster = 'toArray';
+
+  /**
    * Get a default value for a new repository
    *
    * @return mixed