+++ src/Underscore/Types/Object.php
@@ -12,6 +12,12 @@ use \Underscore\Traits\Repository;
 class Object extends Repository
 {
   /**
+   * The method used to convert new subjects
+   * @var string
+   */
+  protected $typecaster = 'toObject';
+
+  /**
    * Get a default value for a new repository
    *
    * @return mixed