+++ src/Underscore/Types/String.php
@@ -10,4 +10,10 @@ use \Underscore\Traits\Repository;
 
 class String extends Repository
 {
+  /**
+   * The method used to convert new subjects
+   * @var string
+   */
+  protected $typecaster = 'toString';
+
 }